Se acomodó problema que eliminaba imágen perfil.png

+++ Nuevo/src/Logica/ControladorProveedores.java
@@ -563,7 +563,9 @@ public class ControladorProveedores implements IControladorProveedores{
         File[]imagenes = fichero.listFiles();
         
         for(int i = 0; i < imagenes.length; i++){
-            imagenes[i].delete();
+            if(!imagenes[i].getName().equalsIgnoreCase("perfil.png"))            
+                imagenes[i].delete();
+            //System.out.println(imagenes[i].getName());
         }
     }
